+++ janito/providers/atria/__init__.py
@@ -1,0 +1,1 @@
+"""Atria provider configuration package."""